Moved Footer to own component. Google fonts added.
